Move GET_OPTIONS_INCLUDE_ALL to test-util.
